feat: adds toggle to profile dropdown for 4h forecast
