+++ apps/nowcasting-app/components/charts/gsp-pv-remix-chart/index.tsx
@@ -23,14 +23,13 @@ const GspPvRemixChart: FC<{
   resetTime: () => void;
   visibleLines: string[];
 }> = ({ gspId, selectedTime, close, setTimeOfInterest, timeNow, resetTime, visibleLines }) => {
-  // add gsp4HourData back in here when working on global state variable for show4hView
-  const { errors, fcAll, pvRealDataAfter, pvRealDataIn } = useGetGspData(gspId);
+  const { errors, fcAll, pvRealDataAfter, pvRealDataIn, gsp4HourData } = useGetGspData(gspId);
   const gspData = fcAll?.forecasts.find((fc) => fc.location.gspId === gspId);
   const gspForecastData = gspData?.forecastValues;
   const gspInfo = gspData?.location;
   const chartData = useFormatChartData({
     forecastData: gspForecastData,
-    // fourHourData: gsp4HourData,
+    fourHourData: gsp4HourData,
     pvRealDayInData: pvRealDataIn,
     pvRealDayAfterData: pvRealDataAfter,
     timeTrigger: selectedTime
